requirements.txt
Django==1.11.6
djangorestframework==3.7.0
Markdown==2.6.9
django-filter==1.0.4
Pillow==4.3.0